requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
influxdb-client==1.31.0
numpy==1.23.2
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
Rx==3.2.0
six==1.16.0
urllib3==1.26.11
